More package restructuring
