+++ .settings/.gitignore
@@ -1,0 +1,1 @@
+/org.eclipse.jdt.core.prefs
