+++ src/com/app/calculator/common/ViewConstants.java
@@ -5,7 +5,7 @@ package com.app.calculator.common;
  * @author Miguel Alvarez
  *
  */
-public class CalculatorConstants {
+public class ViewConstants {
 
 	public static int WIDTH = 50;
