+++ src/com/app/calculator/view/button/action/FirstNumberAction.java
@@ -6,7 +6,7 @@ import java.awt.event.ActionListener;
 import javax.swing.JButton;
 import javax.swing.JTextField;
 
-import com.app.calculator.model.CalculatorInputVO;
+import com.app.calculator.view.bean.InputVO;
 
 /**
  * 
@@ -16,7 +16,7 @@ import com.app.calculator.model.CalculatorInputVO;
 public class FirstNumberAction implements ActionListener{
 	
 	private JTextField textField;
-	private CalculatorInputVO inputVo;
+	private InputVO inputVo;
 	private JButton button;
 
 	/**
@@ -25,7 +25,7 @@ public class FirstNumberAction implements ActionListener{
 	 * @param inputVo
 	 * @param button
 	 */
-	public FirstNumberAction(JTextField textField, CalculatorInputVO inputVo, JButton button) {
+	public FirstNumberAction(JTextField textField, InputVO inputVo, JButton button) {
 		super();
 		this.textField = textField;
 		this.inputVo = inputVo;
