+++ src/com/app/calculator/view/button/action/SecondNumberAction.java
@@ -6,8 +6,8 @@ import java.awt.event.ActionListener;
 import javax.swing.JButton;
 import javax.swing.JTextField;
 
-import com.app.calculator.controller.CalculatorController;
-import com.app.calculator.model.CalculatorInputVO;
+import com.app.calculator.controller.ResultController;
+import com.app.calculator.view.bean.InputVO;
 
 /**
  * 
@@ -17,8 +17,8 @@ import com.app.calculator.model.CalculatorInputVO;
 public class SecondNumberAction implements ActionListener{
 
 	private JTextField textField;
-	private CalculatorInputVO inputVo;
-	private CalculatorController controller;
+	private InputVO inputVo;
+	private ResultController controller;
 	private JButton button;
 	private boolean isOperated;
 	
@@ -30,7 +30,7 @@ public class SecondNumberAction implements ActionListener{
 	 * @param button
 	 * @param isOperated
 	 */
-	public SecondNumberAction(JTextField textField, CalculatorInputVO inputVo, CalculatorController controller, JButton button,
+	public SecondNumberAction(JTextField textField, InputVO inputVo, ResultController controller, JButton button,
 			boolean isOperated) {
 		super();
 		this.textField = textField;
